requirements.txt
bs4
case_conversion
clipboard
colour
fuzzywuzzy
html-text
inflect
maybe-else
msoffcrypto-tool
parsedatetime
prettierfier
python-dateutil
regex
requests
simplejson
xlsxwriter
